app/requirements/base.txt
Flask==2.3.2
SQLAlchemy==2.0.18
Flask-SQLAlchemy==3.0.5
Flask-RESTful==0.3.10
marshmallow==3.19.0
Flask-Migrate==4.0.4
psycopg2==2.9.6
Flask-Script==2.0.6
click==8.1/4
python-dotenv==1.0.0
